aider: Added a print statement to display the message "History file used: histfile" when the Python interactive shell is started.
# Aider chat conversation:

USER: How do I dump a message when python interactive is started? I want to see "History file used: histfile"

ASSISTANT: To dump a message when the Python interactive shell is started, you can print the message at the end of the `config/python/startup` file. This message will be displayed each time the Python interactive shell is initialized. Here's the diff to add the print statement:

```diff
--- config/python/startup
+++ config/python/startup
@@ ... @@
 atexit.register(save, h_len, histfile)
+
+print(f"History file used: {histfile}")
```

This change adds a print statement that will output the message "History file used: histfile" to the console, where `histfile` is the variable holding the path to the history file. This message will be displayed every time a new Python interactive shell session is started.